requirements.txt
Django==5.0.7
psycopg2
pydantic
django-environ
django-tasks
inflection==0.2.0
shapely
django-picklefield
django-model-utils
django-reversion
geojson
fiona